% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beltob.R
\name{beltob}
\alias{beltob}
\title{Computing an implicability function from a belief function.}
\usage{
beltob(bel)
}
\arguments{
\item{bel}{: belief function}
}
\value{
b : implicability function
}
\description{
\code{beltob} converts belief function to implicability function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
