% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beltom.R
\name{beltom}
\alias{beltom}
\title{Computing a mass function from a belief function.}
\usage{
beltom(bel)
}
\arguments{
\item{bel}{: belief function}
}
\value{
m : mass function
}
\description{
\code{beltom} converts belief function to mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
