% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beltopl.R
\name{beltopl}
\alias{beltopl}
\title{Computing a plausibility function from a belief function.}
\usage{
beltopl(bel)
}
\arguments{
\item{bel}{: belief function}
}
\value{
m : plausibility function
}
\description{
\code{beltopl} converts belief function to plausibility function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
