% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beltoq.R
\name{beltoq}
\alias{beltoq}
\title{Computing a commonality function from a belief function.}
\usage{
beltoq(bel)
}
\arguments{
\item{bel}{: belief function}
}
\value{
q : commonality function
}
\description{
\code{beltopl} converts belief function to commonality function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
SS. Mutmainah, D. Mercier
}
