% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btobel.R
\name{btobel}
\alias{btobel}
\title{Computing a belief function from an implicability function.}
\usage{
btobel(b)
}
\arguments{
\item{b}{: implicability function}
}
\value{
bel : belief function
}
\description{
\code{btobel} converts implicability function to belief function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
