% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btom.R
\name{btom}
\alias{btom}
\title{Computing a mass function from an implicability function.}
\usage{
btom(b)
}
\arguments{
\item{b}{: implicability function}
}
\value{
m : mass function
}
\description{
\code{btom} converts implicability function to mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
SS. Mutmainah, D. Mercier
}
