% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btopl.R
\name{btopl}
\alias{btopl}
\title{Computing a plausibility function from an implicability function.}
\usage{
btopl(b)
}
\arguments{
\item{b}{: implicability function}
}
\value{
pl : plausibility function
}
\description{
\code{btopl} converts implicability function to plausibility function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
