% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btoq.R
\name{btoq}
\alias{btoq}
\title{Computing a commonality function from an implicability function.}
\usage{
btoq(b)
}
\arguments{
\item{b}{: implicability function}
}
\value{
q : commonality function
}
\description{
\code{btoq} converts implicability function to commonality function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
