% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cftobetpl.R
\name{cftobetpl}
\alias{cftobetpl}
\title{Computing Betpl (normalized contour function) from a contour function which is plausibility on singletons.}
\usage{
cftobetpl(cf)
}
\arguments{
\item{cf}{: contour function}
}
\value{
betpl : normalized contour function
}
\description{
\code{cftobetpl} computes contour function to betpl.
}
\references{
B. R. Cobb and P. P. Shenoy.  On the plausibility transformation method
         for translating belief function models to probability models.
         InternationalJournal of Approximate Reasoning, 41(3):314–330, 2006.
}
\author{
S. Mutmainah, D. Mercier, F. Pichon, S. Hachour
}
