% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClass.R
\name{checkClass}
\alias{checkClass}
\title{Checking if number of class in training set is proper to the number of class of the object in universe}
\usage{
checkClass(m_train, uniqueClass, ytrain)
}
\arguments{
\item{m_train}{a matrix which are mass functions (e.g., EkNN output consisting of mass functions on singletons and the universe).}

\item{uniqueClass}{vector of the unique class in universe}

\item{ytrain}{labels of training set (ground truth)}
}
\value{
a matrix which as mass functions with correct position according to the actual number of classes in universe.
}
\description{
\code{checkClass} returns approprite matrix containing mass function which is output of classifier.
}
\examples{
m_train = matrix(c(0.3,0.6,0.1,0.7,0.2,0.1),nrow=2,byrow=TRUE) ## for 2 objects
uniqueClass = c(1,2,3) ## the unique class in universe
ytrain = c(1,2) ## the ground truth of the 2 objects in training data
checkClass(m_train, uniqueClass, ytrain);
## note that this function is suitable to output of evidential classifier
## such as Evidential KNN, Evidential NN consiting of singletons and the universe.
}
\author{
S. Mutmainah
}
