% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrectEKNN.R
\name{CorrectEKNN}
\alias{CorrectEKNN}
\title{Correct EkNN classifier using beta vector containing parameters minimizing Epl}
\usage{
CorrectEKNN(
  x_train,
  y_train,
  xtuning_train,
  ytuning_train,
  xtuning_test,
  ytuning_test,
  nclass,
  kn = 5
)
}
\arguments{
\item{x_train}{: predictors of data for training classifier}

\item{y_train}{: ground truths of data for training classifier}

\item{xtuning_train}{: predictors of data for training corrections}

\item{ytuning_train}{: ground truthts of data for training corrections}

\item{xtuning_test}{: predictors of data for testing}

\item{ytuning_test}{: ground truths of data for testing}

\item{nclass}{: number of class of object in universe}

\item{kn}{: number of neighbor}
}
\value{
A list with mass functions output by EkNN and two elements of each correction (CD, CR, CN):
  \describe{
  \item{MF}{Outputs of EkNN classifier}
  \item{beta_CD}{parameters of CD learnt from data for training corrections}
  \item{CD}{Corrected mass functions by CD for the test data. The last column corresponds to the mass
  assigned to the whole set of classes. The other ones correspond
  to the mass assigned to each class.}
  \item{beta_CR}{parameters of CR learnt from data for training corrections}
  \item{CR}{Corrected mass functions by CR for the test data. The last column corresponds to the mass
  assigned to the whole set of classes. The other ones correspond
  to the mass assigned to each class.}
  \item{beta_CN}{parameters of CR learnt from data for training corrections}
  \item{CN}{Corrected mass functions by CR for the test data. The last column corresponds to the mass
  assigned to the whole set of classes. The other ones correspond
  to the mass assigned to each class.}
 }
}
\description{
\code{CorrectEKNN} improves outputs of EkNN classifier using beta vector learnt from data for training corrections.
}
\references{
F. Pichon, D. Mercier, E. Lefèvre, and F. Delmotte.
           Proposition and learn-ing of some belief function contextual correction mechanisms.
           InternationalJournal Approximate Reasoning, 72:4–42, 2016.
}
\author{
S. Mutmainah
}
