% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csum.R
\name{csum}
\alias{csum}
\title{Computing conjunctive combination of two mass functions (m1 and m2)}
\usage{
csum(m1, m2)
}
\arguments{
\item{m1}{: mass function from source 1}

\item{m2}{: mass function from source 2}
}
\value{
new mass function resulting from conjunctive rule of combination
}
\description{
\code{csum} computes conjunctive rule of combination m1 and m2
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
