% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen2.R
\name{datagen2}
\alias{datagen2}
\title{Generating synthetic data with 2 classes and 100 instances}
\usage{
datagen2(mu1 = c(2, 3), mu2 = c(3, 4))
}
\arguments{
\item{mu1}{mean of class 1}

\item{mu2}{mean of class 2}
}
\value{
synthetic data.
}
\description{
\code{datagen2} generates synthetic data with 2 classes and 100 instances.
}
\author{
S. Mutmainah
}
