% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerator.R
\name{datagenerator}
\alias{datagenerator}
\title{Generating synthetic data with 2 classes and 100 instances}
\usage{
datagenerator()
}
\value{
data : synthetic data.
}
\description{
\code{datagenerator} generates synthetic data with 5 classes and 5000 instances.
}
\references{
F. Pichon, D. Mercier, E. Lefèvre, and F. Delmotte.
Proposition and learning of some belief function contextual correction mechanisms.
InternationalJournal Approximate Reasoning, 72:4–42, 2016 (Section 8.6 Figure 4).
}
\author{
S. Mutmainah
}
