% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempstersum.R
\name{dempstersum}
\alias{dempstersum}
\title{Computing Dempster rule of combination of two mass functions (m1 and m2)}
\usage{
dempstersum(m1, m2)
}
\arguments{
\item{m1}{: mass function from source 1}

\item{m2}{: mass function from source 2}
}
\value{
new mass function resulting from dempster rule of combination
}
\description{
\code{dempstersum} computes dempster rule of combination which is normalized conjunctive rule of combination
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
