% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.R
\name{disc}
\alias{disc}
\title{Discounting mass functions using single value of discount rate alpha}
\usage{
disc(m, alpha)
}
\arguments{
\item{m}{: mass function that will be discounted}

\item{alpha}{: discount rate}
}
\value{
discounted mass function
}
\description{
\code{disc} discounts mass function using single value of discount rate alpha
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
