% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsum.R
\name{dsum}
\alias{dsum}
\title{Computing disjunctive combination of two mass functions (m1 and m2)}
\usage{
dsum(m1, m2)
}
\arguments{
\item{m1}{: mass function from source 1}

\item{m2}{: mass function from source 2}
}
\value{
new mass function resulting from disjunctive rule of combination
}
\description{
\code{dsum} computes disjunctive rule of combination m1 and m2
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
