% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnormalized.R
\name{mnormalized}
\alias{mnormalized}
\title{Computing normalized mass function.}
\usage{
mnormalized(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
normalized mass function
}
\description{
\code{mnormalized} computes normalized mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
D. Mercier
}
