% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSourceToMCD.R
\name{mSourceToMCD}
\alias{mSourceToMCD}
\title{Computing corrected mass function using Contextual Discounting (CD) from classifier's output.}
\usage{
mSourceToMCD(mSource, params)
}
\arguments{
\item{mSource}{: mass function which is an output of the classifier.}

\item{params}{: a vector containing parameters used to correct the mass function.}
}
\value{
m : corrected mass function.
}
\description{
\code{mSourceToMCD} improves mass functions provided by EkNN classifier using Contextual Discounting (CD).
}
\examples{
mSource <- c(0, 0, 0.9, 0.1) ## output of EkNN (singletons and Omega)
params <- c(0.4, 0.9, 0.1); ## assigned for 3 classes
mSourceToMCD(mSource, params);
## output: 0, 0, 0, 0, 0.324, 0.486, 0.036, 0.154
}
\references{
F. Pichon, D. Mercier, E. Lefèvre, and F. Delmotte.
           Proposition and learn-ing of some belief function contextual correction mechanisms.
           InternationalJournal Approximate Reasoning, 72:4–42, 2016.
}
\author{
S. Mutmainah, D. Mercier, F. Pichon
}
