% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSourceToMCN.R
\name{mSourceToMCN}
\alias{mSourceToMCN}
\title{Computing corrected mass function using Contextual Negating (CN) from classifier's output.}
\usage{
mSourceToMCN(mSource, params)
}
\arguments{
\item{mSource}{: mass function which is an output of the classifier.}

\item{params}{: a vector containing parameters used to correct the mass function.}
}
\value{
m : corrected mass function.
}
\description{
\code{mSourceToMCN} improves mass functions provided by EkNN classifier using Contextual Negating (CN).
}
\examples{
mSource <- c(0, 0, 0.9, 0.1); ## output of EkNN (singletons and Omega)
params <- c(0.4, 0.9, 0.1); ##assigned for 3 classes
mSourceToMCN(mSource, params);
## Output: 0.297, 0.441, 0.081, 0.081, 0.033, 0.049, 0.009, 0.009
}
\references{
F. Pichon, D. Mercier, E. Lefèvre, and F. Delmotte.
           Proposition and learn-ing of some belief function contextual correction mechanisms.
           InternationalJournal Approximate Reasoning, 72:4–42, 2016.
}
\author{
S. Mutmainah, D. Mercier, F. Pichon
}
