% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSourceToMCR.R
\name{mSourceToMCR}
\alias{mSourceToMCR}
\title{Computing corrected mass function using Contextual Reinforcement (CR) from classifier's output.}
\usage{
mSourceToMCR(mSource, params)
}
\arguments{
\item{mSource}{: mass function which is an output of the classifier.}

\item{params}{: a vector containing parameters used to correct the mass function.}
}
\value{
m : corrected mass function.
}
\description{
\code{mSourceToMCR} improves mass functions provided by EkNN classifier using Contextual Reinforcement (CR).
}
\examples{
mSource <- c(0, 0, 0.9, 0.1); ## output of EkNN (singletons and Omega)
params <- c(0.4, 0.9, 0.1); ## assigned for 3 classes
mSourceToMCR(mSource, params);
## Output: 0.8154, 0.0036, 0.0486, 0.0324, 0.0906, 4e-04, 0.0054, 0.0036
}
\references{
F. Pichon, D. Mercier, E. Lefèvre, and F. Delmotte.
           Proposition and learn-ing of some belief function contextual correction mechanisms.
           InternationalJournal Approximate Reasoning, 72:4–42, 2016.
}
\author{
S. Mutmainah, D. Mercier, F. Pichon
}
