% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtob.R
\name{mtob}
\alias{mtob}
\title{Computing an implicability function from a mass function.}
\usage{
mtob(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
b : implicability function
}
\description{
\code{mtob} converts mass function to implicability function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
