% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtobel.R
\name{mtobel}
\alias{mtobel}
\title{Computing a belief function from a mass function.}
\usage{
mtobel(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
bel : belief function
}
\description{
\code{mtobel} converts mass function to belief function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
