% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtobetp.R
\name{mtobetp}
\alias{mtobetp}
\title{Computing pignistic probability (betp) from a mass function.}
\usage{
mtobetp(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
betp : pignistic probability
}
\description{
\code{mtobetp} computes pignistic probability (betp) from mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
