% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtobetpl.R
\name{mtobetpl}
\alias{mtobetpl}
\title{Computing normalized contour function cf (betpl) from a mass function.}
\usage{
mtobetpl(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
betpl : normalized contour function.
}
\description{
\code{mtobetpl} computes betpl from mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
