% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtocf.R
\name{mtocf}
\alias{mtocf}
\title{Computing contour function from a mass function.}
\usage{
mtocf(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
cf : contour function
}
\description{
\code{mtocf} converts mass function to contour function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
