% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtoM.R
\name{mtoM}
\alias{mtoM}
\title{Making a proper matrix of mass functions on the whole frame}
\usage{
mtoM(m)
}
\arguments{
\item{m}{: mass function (with 2^Omega elements, where Omega is the frame of discernment)}
}
\value{
M : matrix of mass functions on the whole frame
}
\description{
\code{mtoM} Return a matrix of mass functions on the whole frame
from a matrix of mass functions given on singletons and the universe
}
\author{
D. Mercier
}
