% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtopl.R
\name{mtopl}
\alias{mtopl}
\title{Computing a plausibility function from a mass function.}
\usage{
mtopl(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
pl : plausibility function
}
\description{
\code{mtopl} converts mass function to plausibility function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
