% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtoq.R
\name{mtoq}
\alias{mtoq}
\title{Computing a commonality function from a mass function.}
\usage{
mtoq(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
q : commonality function
}
\description{
\code{mtoq} converts mass function to commonality function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
