% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtoq0.R
\name{mtoq0}
\alias{mtoq0}
\title{Computing a 0-commonality function from a mass function.}
\usage{
mtoq0(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
q0 : 0-commonality function
}
\description{
\code{mtoq0} converts mass function to 0-commonality function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
 @import pracma
}
\author{
S. Mutmainah, F. Pichon
}
