% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtosetmax.R
\name{mtosetmax}
\alias{mtosetmax}
\title{Computing a set of prediction using the maximality criterion.}
\usage{
mtosetmax(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
out_vec : output vector which is a set of the prediction
}
\description{
\code{mtosetmax} computes mass function to make a set of prediction using the maximality criterion.
}
\examples{
m<-c(.1,.2,.5,.3); mtosetmax(m);
m<-c(.1,0,.5,.4); mtosetmax(m);
}
\references{
Denoeux 2019 Decision-making with belief functions: A review (Section 5.1)

Ma and Denoeux 2021 Partial classification in the belief function framework (Section 3.4)
}
\author{
D. Mercier
}
