% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtosetsd.R
\name{mtosetsd}
\alias{mtosetsd}
\title{Computing a set of partial decision using strong dominance criterion.}
\usage{
mtosetsd(m)
}
\arguments{
\item{m}{: mass function}
}
\value{
out_vec : output vector which is a set of partial decision.
}
\description{
\code{mtosetsd} computes mass function to make a set of partial decision using strong dominance criterion so called interval dominance.
}
\examples{
m<-c(.1,.2,.5,.3); mtosetsd(m);
m<-c(.1,0,.5,.4); mtosetsd(m);
}
\references{
L. Ma, T. Denoeux. Making set-valued predictions in evidential classification:
      A comparison of different approach. International Symposium in Imprecise Probability: Theories and Application
      ISIPTA 2019, 3-6 July 2019, Volume 103 of Proceding of Machine Learning Research, p. 276-285.

T. Denœux. Decision-making with belief functions: A review.
      InternationalJournal of Approximate Reasoning, 109:87–110, 2019.
}
\author{
S. Mutmainah, S. Hachour, F. Pichon, D. Mercier
}
