% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plSourceToplCN.R
\name{plSourceToplCN}
\alias{plSourceToplCN}
\title{Computing corrected pl from mass functions as source outputs}
\usage{
plSourceToplCN(mSource, beta)
}
\arguments{
\item{mSource}{: mass functions as source outputs.}

\item{beta}{: reliability vector of contextual negating (CN).}
}
\value{
pl : Corrected output.
}
\description{
\code{plSourceToplCN} computes corrected pl from mass functions as source outputs.
}
\examples{
mSource <- c(0, 0, 0.9, 0.1) ##output of EkNN (singletons and Omega)
beta <- c(0.4, 0.9, 0.1) ##reliability vector assigned for 3 classes
## return corrected pl by CN: 0.58, 0.18, 0.1
}
\references{
F. Pichon, D. Mercier and E. Lefevre, F. Delmotte.
          Proposition and learning of somebelief function contextual correction mechanims.
          International Journal of Approximate Reasoning, 72:4-42, 2016.
}
\author{
S. Mutmainah
}
