% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltob.R
\name{pltob}
\alias{pltob}
\title{Computing an implicability  function from a plausibility function.}
\usage{
pltob(pl)
}
\arguments{
\item{pl}{: plausibility function}
}
\value{
b : implicability function
}
\description{
\code{pltob} converts plausibility function to implicability function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
