% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltobel.R
\name{pltobel}
\alias{pltobel}
\title{Computing a belief function from a plausibility function.}
\usage{
pltobel(pl)
}
\arguments{
\item{pl}{: plausibility function}
}
\value{
bel : belief function
}
\description{
\code{pltobel} converts plausibility function to belief function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
