% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltocf.R
\name{pltocf}
\alias{pltocf}
\title{Computing a contour function from a plausibility function.}
\usage{
pltocf(pl)
}
\arguments{
\item{pl}{: plausibility function}
}
\value{
cf : contour function
}
\description{
\code{pltocf} converts plausibility function to contour function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
