% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltom.R
\name{pltom}
\alias{pltom}
\title{Computing a mass function from a plausibility function.}
\usage{
pltom(pl)
}
\arguments{
\item{pl}{: plausibility function}
}
\value{
m : mass function
}
\description{
\code{pltom} converts plausibility function to mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
