% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pltoq.R
\name{pltoq}
\alias{pltoq}
\title{Computing a commonality function from a plausibility function.}
\usage{
pltoq(pl)
}
\arguments{
\item{pl}{: plausibility function}
}
\value{
q : commonality function
}
\description{
\code{pltoq} converts plausibility function to commonality function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
