% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q0tom.R
\name{q0tom}
\alias{q0tom}
\title{Computing a mass function from a 0-commonality function.}
\usage{
q0tom(q0)
}
\arguments{
\item{q0}{: 0-commonality function}
}
\value{
m : mass function
}
\description{
\code{q0tom} converts 0-commonality function to mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, F. Pichon
}
