% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtob.R
\name{qtob}
\alias{qtob}
\title{Computing an implicability function from a commonality function.}
\usage{
qtob(q)
}
\arguments{
\item{q}{: commonality function}
}
\value{
b : implicability function
}
\description{
\code{qtob} converts commonality function to implicability function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
