% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtobel.R
\name{qtobel}
\alias{qtobel}
\title{Computing a belief function from a commonality function.}
\usage{
qtobel(q)
}
\arguments{
\item{q}{: commonality function}
}
\value{
bel : belief function
}
\description{
\code{qtobel} converts commonality function to belief function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
