% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtom.R
\name{qtom}
\alias{qtom}
\title{Computing a mass function from a commonality function.}
\usage{
qtom(q)
}
\arguments{
\item{q}{: commonality function}
}
\value{
m : mass function
}
\description{
\code{qtom} converts commonality function to mass function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.
}
\author{
S. Mutmainah, D. Mercier
}
