% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtopl.R
\name{qtopl}
\alias{qtopl}
\title{Computing a plausibility function from a commonality function.}
\usage{
qtopl(q)
}
\arguments{
\item{q}{: commonality function}
}
\value{
pl : plausibility function
}
\description{
\code{qtopl} converts commonality function to plausibility function.
}
\references{
P. Smets and R. Kennes. The Transferable Belief Model.
         Artificial Intelligence, 66:191–243, 1994.

P. Smets. The application of the matrix calculus to belief functions.
        Interna-tional Journal of Approximate Reasoning, 31(1-2):1–30, 2002.
}
\author{
S. Mutmainah, D. Mercier
}
