% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reward_U80.R
\name{reward_U80}
\alias{reward_U80}
\title{Computing performace of classifier outputting partial decision (imprecise output)
using utility-discounted accuracies U80.}
\usage{
reward_U80(x)
}
\arguments{
\item{x}{: tibble of partial predictions which are outputs of credal classifier.}
}
\value{
sum of rewards of correct partial prediction of the classifier
}
\description{
\code{reward_U80} computes performance of classifier outputting partial outputs using U80.
}
\references{
G. Corani, M. Zafallon and D. D. Maua. Evaluating credal classifiers by
          utility discounted predictive accuracy.
          International Journal Approximation Reasoning, 53(8): 1282-1301, 2012
}
\author{
S. Mutmainah
}
