% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmtotv.R
\name{tmtotv}
\alias{tmtotv}
\title{Making the truth given as a vector from the truth given as a matrix.}
\usage{
tmtotv(tm)
}
\arguments{
\item{tm}{: true class of each object given as a matrix such that:
tm[i,j] = 0 iff object i is not of class j
tm[i,j] = 1 iff object i is of class j}
}
\value{
tv: true class of each object given as a vector:
tv[i]=class of object i
}
\description{
\code{tmtotv} gives the truth given as a vector from the truth given as a matrix.
}
\examples{
tm <- matrix(c(1,0,0, 0,1,0, 1,0,0, 0,0,1),nrow = 4, byrow = TRUE);
tmtotv(tm);
## returns 1 2 1 3
}
\author{
S.Mutmainah, D. Mercier
}
