% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningCD_rev2.R
\name{tuningCD_rev2}
\alias{tuningCD_rev2}
\title{Learning discount rate minimizing Epl by contextual discounting (CD)}
\usage{
tuningCD_rev2(data, t)
}
\arguments{
\item{data}{: data from an information source, e.g., sensor, classifier. The matrix consists of lines containing mass functions}

\item{t}{: ground truth}
}
\value{
A list with three elements:
  \describe{
  \item{alpha}{Computing contextual discount rate alpha minimizing the Euclidean distance between corrected contour functions
             (the discounted contour function) and the truth t using Equation 33 of [1]
             (equiv Equation 47 of [2] given for computing beta = 1-alpha)}
  \item{beta}{1-alpha}
  \item{mindist}{Minimum Euclidean distance obtained.}
 }
}
\description{
\code{tuningCD_rev2} optimizes reliability vector by minimizing Epl using
Equation 33 of reference [1] (equivalent of Equation 47 of [2] given for computing beta=1-alpha) with
Least square problem by pseudo inverse.
}
\examples{
## with 4 objects (Data from sensor 1 of reference [3] Elouedi 2004):
data1=matrix(c(0,0,0,0,0.5,0,0.3,0.2, 0,0,0.5,0,0.2,0,0,0.3,0,0,0.4,0,0,0.6,
0,0,0,0,0,0,0,0.6,0.4,0),nrow=4,byrow = TRUE)
t <- matrix(c(1,0,0, 0,1,0, 1,0,0, 0,0,1), nrow = 4, byrow = TRUE)
tuningCD_rev2(data1,t);
## returns
 ##     $alpha: 0.2413793 -0.6831683 -1.1951220 ## discount rates alpha should
 ##     be [0,1], while by using psedo-inverse the result may be different.
 ##     $beta: 0.7586207 1.6831683 2.1951220 ## beta = 1-alpha
 ##     $mindist: 2.328521
}
\references{
[1] D. Mercier, B. Quost, T. Denoeux
      Refined modeling of sensor reliability in the belief function framework using contextual discounting,
      Information Fusion, Vol. 9, Issue 2, pp 246-258, April 2008.

  [2] F. Pichon, D. Mercier, É. Lefèvre, F. Delmotte
      Proposition and learning of some belief function contextual correction mechanisms,
      International Journal of Approximate Reasoning, Vol. 72, pp 4-42, May 2016

  [3] Z. Elouedi, F. Mellouli, P. Smets
      Assessing sensor reliability for multisensor data fusion within the transferable belief model.
      IEEE Trans Syst Man Cybern B Cybern. 2004 Feb;34(1):782-7.
 @import pracma
}
\author{
S.Mutmainah, D. Mercier, F. Pichon, S. Hachour
}
