% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningCN.R
\name{tuningCN}
\alias{tuningCN}
\title{Learning discount rate minimizing Epl by contextual negating (CN)}
\usage{
tuningCN(data, t)
}
\arguments{
\item{data}{: data from an information source, e.g., sensor, classifier. The matrix consists of lines containing mass functions}

\item{t}{: ground truth}
}
\value{
A list with three elements:
  \describe{
  \item{alpha}{Computing contextual discount rate alpha minimizing the Euclidean distance between corrected contour functions
             (the discounted contour function) and the truth t using Equation 33 of [1]
             (equiv Equation 47 of [2] given for computing beta = 1-alpha)}
  \item{beta}{1-alpha}
  \item{mindist}{Minimum Euclidean distance obtained.}
 }
}
\description{
\code{tuningCN} optimizes reliability vector by minimizing Epl using
Equation 33 of reference [1] (equivalent of Equation 47 of [2] given for computing beta=1-alpha).
}
\examples{
## with 4 objects (Data from sensor 1 of reference [3] Elouedi 2004):
data1=matrix(c(0,0,0,0,0.5,0,0.3,0.2, 0,0,0.5,0,0.2,0,0,0.3,0,0,0.4,0,0,0.6,
0,0,0,0,0,0,0,0.6,0.4,0),nrow=4,byrow = TRUE);
t <- matrix(c(1,0,0, 0,1,0, 1,0,0, 0,0,1), nrow = 4, byrow = TRUE)
tuningCN(data1,t);
## returns
 ##     $alpha: 0.6666667 0.0000000 0.5490196
 ##     $beta: 0.3333333 1.0000000 0.4509804
 ##     $mindist: 2.588431
}
\references{
[1] D. Mercier, B. Quost, T. Denoeux
      Refined modeling of sensor reliability in the belief function framework using contextual discounting,
      Information Fusion, Vol. 9, Issue 2, pp 246-258, April 2008.

  [2] F. Pichon, D. Mercier, É. Lefèvre, F. Delmotte
      Proposition and learning of some belief function contextual correction mechanisms,
      International Journal of Approximate Reasoning, Vol. 72, pp 4-42, May 2016

  [3] Z. Elouedi, F. Mellouli, P. Smets
      Assessing sensor reliability for multisensor data fusion within the transferable belief model.
      IEEE Trans Syst Man Cybern B Cybern. 2004 Feb;34(1):782-7.
}
\author{
S.Mutmainah, D. Mercier, F. Pichon, S. Hachour
}
