% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvtotm.R
\name{tvtotm}
\alias{tvtotm}
\title{Making the truth given as a matrix from the truth given as a vector}
\usage{
tvtotm(tv, nclass)
}
\arguments{
\item{tv}{true class of each object given as a vector: tv[i]=class of object i}

\item{nclass}{number of classes (some being possibly not in the vector)}
}
\value{
tm: true class of each object given as a matrix such that:
tm[i,j] = 0 iff object i is not of class j
tm[i,j] = 1 iff object i is of class j
}
\description{
\code{tmtotv} gives the truth given as a matrix from the truth given as a vector
}
\examples{
  tv <- c(1,2,1,3);
  nclass <- 3;
  tvtotm(tv,nclass); ## change a vector to a matrix
}
\author{
S.Mutmainah, D. Mercier, F. Pichon, S. Hachour
}
